Even more base/logging.h removal and fixes
