+++ GPU/Directx9/ShaderManagerDX9.cpp
@@ -23,7 +23,6 @@
 #include <map>
 
 #include "gfx/d3d9_shader.h"
-#include "base/logging.h"  // For OutputDebugStringUTF8
 #include "base/stringutil.h"
 #include "i18n/i18n.h"
 #include "math/lin/matrix4x4.h"
