+++ ext/native/base/stringutil.cpp
@@ -21,27 +21,6 @@
 #include "base/buffer.h"
 #include "base/stringutil.h"
 
-#ifdef _WIN32
-
-void OutputDebugStringUTF8(const char *p) {
-	wchar_t temp[4096];
-
-	int len = std::min(4095, (int)strlen(p));
-	int size = (int)MultiByteToWideChar(CP_UTF8, 0, p, len, NULL, 0);
-	MultiByteToWideChar(CP_UTF8, 0, p, len, temp, size);
-	temp[size] = 0;
-
-	OutputDebugString(temp);
-}
-
-#else
-
-void OutputDebugStringUTF8(const char *p) {
-	INFO_LOG(SYSTEM, "%s", p);
-}
-
-#endif
-
 std::string LineNumberString(const std::string &str) {
 	std::stringstream input(str);
 	std::stringstream output;
