+++ ext/native/json/json_reader.h
@@ -1,9 +1,9 @@
+#include <cstring>
 #include <string>
 #include <vector>
 
 #include "base/basictypes.h"
 #include "ext/gason/gason.h"
-
 #include "Common/Log.h"
 
 namespace json {
