+++ headless/WindowsHeadlessHost.cpp
@@ -15,9 +15,8 @@
 // Official git repository and contact information can be found at
 // https://github.com/hrydgard/ppsspp and http://www.ppsspp.org/.
 
-#include <stdio.h>
+#include <cstdio>
 
-#include "base/logging.h"  // For OutputDebugStringUTF8
 #include "headless/WindowsHeadlessHost.h"
 
 #include "Common/Log.h"
